Extend CSVTournamentFactory API to deal with charsets
Currently the only real benefit is that the IYPT2012 example is
displayed correctly because the platform-default encoding is ignored
when readinf classpath resources (that _are_ UTF-8 encoded).

The API now allows for choosing a charset which could save the situation
when working with e.g. UTF-8 encoded files on Windows. However, for this
to be really usable it would require some UI support.